xai_agent_research/requirements.txt
torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
tokenizers==0.22.2
cohere===5.20.1
pillow==10.0.0
peft==0.9.0
timm==0.9.12
rfdetr==1.2.1
supervision==0.26.1
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
google-generativeai==0.7.2
langgraph==0.2.16
llama-index==0.10.30
llama-index-embeddings-huggingface==0.2.0
llama-index-postprocessor-cohere-rerank==0.1.4
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.7.1
opencv-python==4.9.0.80
numpy==1.24.3